requirements.txt
# ML & Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
shap>=0.44.0
joblib>=1.3.0

# API Backend
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0

# AWS Services
boto3>=1.34.0
python-dotenv>=1.0.0
watchtower>=3.0.0

# GitHub Integration
requests>=2.31.0
